Collection that tracks observable sequences
TrackingCollection gets items from an observable sequence and updates
its contents in such a way that two updates to the same object (as
defined by an Equals call) will result in one object on the list being
updated (as opposed to having two different instances of the object
added to the list).

This allows us to grab a bunch of objects from cache, load a collection
with them, and then request updated versions of the objects from the
server, which when returned will update the collection in place (and the
UI with it), seamlessly.

The collection supports ordering and filtering, and filtering includes
enough information to allow limiting the number of items to return (and
display). It can be resorted and refiltered in place.